+++ src/common/osc_file_parser.js
@@ -1,4 +1,4 @@
-const { DOMParser, XMLSerializer } = require('xmldom')
+const { DOMParser, XMLSerializer } = require('@xmldom/xmldom')
 const xmlToJSON = require('xmltojson')
 const _fromPairs = require('lodash.frompairs')
 const _find = require('lodash.find')
@@ -98,7 +98,7 @@ const OSCFileParser = {
    */
   explode: async function(xmlString) {
     // Note that "nodes" here refer to XML nodes, not OSM nodes
-    const doc = new DOMParser().parseFromString(xmlString)
+    const doc = new DOMParser().parseFromString(xmlString, 'text/xml')
     const serializer = new XMLSerializer()
     const separateChanges = []
 
@@ -126,7 +126,7 @@ const OSCFileParser = {
   },
 
   josmToOSC: function(xmlString) {
-    const doc = new DOMParser().parseFromString(xmlString)
+    const doc = new DOMParser().parseFromString(xmlString, 'text/xml')
     const serializer = new XMLSerializer()
     const changeOperations = []